A_1_MultiAgent_Research_Assistant_with_MCP_Integration/requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-groq

mcp
groq
ddgs

psycopg2-binary

python-dotenv

streamlit

django
djangorestframework

pydantic
typing-extensions

ipykernel
jupyter

pgvector
openai
pandas